% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanRbp-package.R
\docType{package}
\name{cleanRbp-package}
\alias{cleanRbp}
\alias{cleanRbp-package}
\title{cleanRbp: Cleaner Blood Pressure Data}
\description{
Blood pressure (BP) is routinely measured in clinical practice.
    BP may also be monitored over 24-hours using ambulatory monitoring 
    devices or home BP monitoring devices. There are a number of variables
    derived from these types of BP data that measure overall BP burden and
    BP variability. Additionally, these variables may be of interest during
    particular periods of the day (e.g., while a person is asleep). The 
    cleanRbp package exists to simplify data processing for these analyses.
}
\author{
\strong{Maintainer}: Byron Jaeger \email{bcjaeger@uab.edu} (\href{https://orcid.org/0000-0001-7399-2299}{ORCID})

}
\keyword{internal}
