% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_arv.R
\name{cmp_arv}
\alias{cmp_arv}
\title{Compute average real variability}
\usage{
cmp_arv(time, value, phase = NULL)
}
\arguments{
\item{time}{numeric vector indicating time of measurement}

\item{value}{numeric vector indicating measurement reading.}

\item{phase}{numeric vector indicating the phase of measurement.}
}
\description{
Compute average real variability
}
\examples{
# Example 1 -----

time <- c(1,2,3,4)
value <- c(0,1,3,6)
# time differences = 1, 1, 1
# value differences = 1, 2, 3
# ARV = (1/3) * (1*1 + 1*2 + 1*3) = 1/3 * 6 = 2
cmp_arv(time, value)

# Example 2 -----

time <- c(1,3,4,10)
value <- c(0,1,3,6)
# time differences = 2, 1, 6
# value differences = 1, 2, 3
# ARV = (1/9) * (2*1 + 1*2 + 6*3) = 1/9 * 22 = 2.444
cmp_arv(time, value)

# Example 3 (same as 2 but with negative values) -----

time <- c(1,3,4,10)
value <- c(0,1,3,6) * -1
# time differences = 2, 1, 6
# value differences = 1, 2, 3
# ARV = (1/9) * (2*1 + 1*2 + 6*3) = 1/9 * 22 = 2.444
cmp_arv(time, value)

# Example 4 (missing values) -----

time <- c(1, 2, 4, 6)
value <- c(0, 1, NA_real_, 2)
# time differences = 1, 4
# value differences = 1, 1
# ARV = (1/5) * (1*1 + 4*1) = 5/5 = 1
cmp_arv(time, value)

# Example 5 (phases) -----

time <- c(1,3,  4,10)

value <- c(0,1,3,6)

phase <- c(1,1,2,2)

# time differences = 2, 6
# value differences = 1, 3
# ARV = {2 * [(1/2) * (2*1)] + 6 * [(1/6) * (6*3)]} / 8
#     = {2*1 + 6*3} / 8
#     = 20 / 8, i.e., 2.5

cmp_arv(time, value, phase)


# Example 6 (phases) -----

time <- c(1,3,12,  4,10,13)
value <- c(0,1,5,  3,6,1)
phase <- c(1,1,1,  2,2,2)

# time differences = {2, 9}, {6, 3} - totals of 11 and 9
# value differences = {1, 4}, {3, -5}
# ARV = (11 * ((1/11) * (2*1 + 9*4)) + 9 * ((1/9) * (6*3 +3*5))) / 20 = 3.55

cmp_arv(time, value, phase)
}
