% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sdim.R
\name{cmp_sdim_coefs}
\alias{cmp_sdim_coefs}
\alias{cmp_sdim_values}
\title{Standard deviation independent of mean}
\usage{
cmp_sdim_coefs(bp_mean, bp_sd)

cmp_sdim_values(bp_mean, bp_sd)
}
\arguments{
\item{bp_mean}{a numeric vector of mean blood pressure values.}

\item{bp_sd}{a numeric vector of standard deviations corresponding to the
mean blood pressure values.}
}
\value{
\link{cmp_sdim_coefs} returns the regression coefficients from
a linear model fitted to the log-transformed values of \code{bp_mean}
and \code{bp_sd} (i.e., \code{coef(lm(log_sd ~ log_mean))}). \link{cmp_sdim_values}
uses the regression coefficients form \link{cmp_sdim_coefs} to compute
the standard deviation of blood pressure independent from the mean.
}
\description{
Standard deviation independent of mean
}
\note{
SDIM (standard deviation independent of the mean) is cohort-specific.
So, if you remove a participant from the study, it will result in a
new value of SDIM for everyone else.
}
\examples{
mean_bps = runif(n = 100, min = 80, max = 190)
sd_bps   = runif(n = 100, min = 1, max = 3)
cmp_sdim_coefs(mean_bps, sd_bps)
cmp_sdim_values(mean_bps, sd_bps)
}
