% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_hypertension.R
\name{htn_classify}
\alias{htn_classify}
\title{Classify hypertension}
\usage{
htn_classify(
  sbp,
  dbp,
  sbp_thresh = 130,
  dbp_thresh = 80,
  strictly_greater = FALSE,
  labels = c("normotension", "hypertension")
)
}
\arguments{
\item{sbp}{(numeric) -- systolic blood pressure values}

\item{dbp}{(numeric) -- diastolic blood pressure values}

\item{sbp_thresh}{(numeric) -- a threshold for systolic blood pressure.
A value of \code{sbp} exceeding \code{sbp_thresh} indicates hypertension}

\item{dbp_thresh}{(numeric) -- a threshold for diastolic blood pressure.
A value of \code{dbp} exceeding \code{dbp_thresh} indicates hypertension}

\item{strictly_greater}{(logical) -- Should hypertension only be indicated
if a \code{sbp} or \code{dbp} value is strictly greater (i.e., not equal to) the
corresponding threshold value? (default is \code{FALSE})}

\item{labels}{(character) -- How should normotensive and hypertensive
observations be labeled? Default is "normo" and "hyper" for normotensive
and hypertensive blood pressure readings, respectively.}
}
\value{
a factor with two levels, one indicating normotension and the
other indicating hypertension, created using the given systolic
and diastolic blood pressure thresholds.
}
\description{
Classify hypertension
}
\examples{

sbp <- c(120, 130, 140)
dbp <- c(70, 80, 90)

htn_classify(sbp, dbp)

}
