% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abpm_mark_extremes.R
\name{mark_max}
\alias{mark_max}
\alias{mark_min}
\title{ABPM peaks/troughs}
\usage{
mark_max(x, neighborhood_size = 1)

mark_min(x, neighborhood_size = 1)
}
\arguments{
\item{x}{(numeric) ambulatory monitoring values}

\item{neighborhood_size}{(numeric) the number of points to include
around the maximum or minimum value. If the maximum or minimum
occurs on a boundary, then only neighbors within the boundary
will be marked.}
}
\value{
a numeric vector the same length as \code{x} with values of
0 or 1. A value of 1 indicates that the corresponding value of
\code{x} is inside of the peak or trough.
}
\description{
ABPM peaks/troughs
}
\note{
If there are ties, the value that appears
}
\examples{

x <- c(110, 125, 120, NA, 110)
mark_min(x, neighborhood_size = 0)
mark_max(x, neighborhood_size = 0)
mark_min(x, neighborhood_size = 1)
mark_max(x, neighborhood_size = 1)

}
